ecore_cocoa: fix windows initial rendering
We need to refresh the Cocoa's content view when showing the window.
In some particular cases (including the elementary_test initial window),
the content view is redrawn because of external events (the entry
animator). When no action at all is performed, the window and the
content view are resized... nice... but not drew...

Such a simple fix for this hell of a bug!

Fixes T5210