bump minimum version requirement of freetype2 to 16.2.10 which equals release 2.5.0.1
For this one I feel like I earned an archaeologist medal. :-)

This is related to the commit raster made in aa92cddb to accommodate for older
freetype version. The TT_INTERPRETER_VERSION_35 define was actually introduced
in the release 2.5.0 which was released in June 2013 which means it had almost
4 years to get picked up. I would consider this safe to be a minimum dependency
for current efl versions.

The trick here is that the release version does not match the library version
spit out by pkg-config. 16.2.10 equals the freetype 2.5.0.1 release in this case.
Better make a note about it in configure for the next poor soul trying to
understand this.

Thanks again to llelectronics for reporting this.
Fix T5437
@fix