mono: Update function pointer support
Removed the IntPtr data and Free function, leaving just a simple
delegate on the C# side.

When implementing virtual functions receiving the function pointers,
they are wrapped into a lambda and passed to C#.

One issue is that the free func is currently not being handled in these
virtual calls.