efl_debugd: don't compile this on macOS
Well, the build is broken again on macOS :(
efl_debugd is full of platform-specific code: network and epoll.
To make it compile (not work as expected, just compile), we would need
to comment out most of efl_debugd's code. So instead of having a
crippled, useless binary, just don't compile it on macOS.

This unbreaks the build for macOS.