slider: Fix slider legacy API
See 4e79dd0f02811a013b461d4839f14a48ee9b1c2f

That patch was absurd. Do not change the use of a legacy stable
API when you change an EO API. If you need to do that then there
is very clearly a problem in the patch.

This reverts the test case to use the legacy API (which in turn
calls the EO API anyway so both are tested).

Fixes T5587